Rename: essentialInfoParser user.parser로 이동
