+++ webapp/src/service/auth.parser.js
@@ -14,28 +14,3 @@ export const userInfoParser = (userInfoRaw) => {
   const { id, image, name } = userInfoRaw;
   return { id, nickname: name, profileImg: image };
 };
-
-export const essentialInfoParser = (essentialInfoRawData) => {
-  const {
-    introduction,
-    hopeSession,
-    profileImage,
-    job,
-    nickname,
-    portfolio,
-    slogan,
-    techSkills,
-    belongTeam,
-  } = essentialInfoRawData;
-  const paresedTechSkills = skillStackParserToIds(techSkills);
-  return {
-    content: introduction,
-    hope_session: hopeSession,
-    image: profileImage,
-    job,
-    name: nickname,
-    portfolio,
-    slogan,
-    skills: paresedTechSkills,
-  };
-};
